@@ -15,6 +15,8 @@ import { TokenInterceptorService } from './services/token-interceptor.service';
 import { ListCatsComponent } from './list-cats/list-cats.component';
 import { DetailsCatComponent } from './details-cat/details-cat.component';
 import { EditCatComponent } from './edit-cat/edit-cat.component';
+import { ErrorInterceptorService } from './services/error-interceptor.service';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
   declarations: [
@@ -30,7 +32,8 @@ import { EditCatComponent } from './edit-cat/edit-cat.component';
     BrowserModule,
     AppRoutingModule,
     ReactiveFormsModule,
-    HttpClientModule
+    HttpClientModule,
+    BrowserAnimationsModule,
   ],
   providers: [
     AuthService, 
@@ -39,6 +42,11 @@ import { EditCatComponent } from './edit-cat/edit-cat.component';
     {
       provide: HTTP_INTERCEPTORS,
       useClass: TokenInterceptorService,
+      multi: true
+    },
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: ErrorInterceptorService,
       multi: true
     }
 ],
